requirements.txt
flickrapi>=2.4
requests>=2.31
requests_oauthlib>=1.3
python-dotenv>=1.0